Add MVP
Sort-of working with three suns and a planet, want to scale up stuff to
realistic sizes and make the planet super dense but walkable, so that
a character can walk on the surface and observe the suns' movement.

Bonus is enablin fast-forwarding time, adding a starry sky backdrop to
see relative movement, making the suns light sources, estimating surface
temperatures on the planet, etc.